[REFACTOR] #115 - 신고 activity 뷰모델을 분리
- 신고 관련 로직만 갖고 있는 reportViewModel로 수정
